Added All the weapons bar shotgun; added melee; added reloading & dropping weapons; adding ranges

+++ Assets/Scripts/Enemy/EnemyHealthTest.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public class EnemyHealthTest : MonoBehaviour
 {
-    public int currentHealth;
+    public float currentHealth;
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
         currentHealth -= amount;
         if(currentHealth <= 0)
